refactor(retriever): :fire: move Retriever methods closer to type definition
