+++ pkg/controller/servicebindingrequest/retriever.go
@@ -5,9 +5,13 @@ import (
 	"strings"
 
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/dynamic"
 
+	"github.com/imdario/mergo"
 	"github.com/redhat-developer/service-binding-operator/pkg/log"
+
+	"github.com/redhat-developer/service-binding-operator/pkg/apis/apps/v1alpha1"
 )
 
 // Retriever reads all data referred in plan instance, and store in a secret.
@@ -20,6 +24,71 @@ type Retriever struct {
 	VolumeKeys    []string                     // list of keys found
 	bindingPrefix string                       // prefix for variable names
 	cache         map[string]interface{}       // store visited paths
+}
+
+// GetEnvVars returns the data read from related resources (see ReadBindableResourcesData and
+// ReadCRDDescriptionData).
+func (r *Retriever) GetEnvVars() (map[string][]byte, error) {
+	envVarCtx := make(map[string]interface{})
+
+	for _, resource := range r.plan.GetServiceContexts() {
+		// contribute values extracted from the service related resources
+		err := mergo.Merge(&envVarCtx, resource.EnvVars, mergo.WithAppendSlice, mergo.WithOverride)
+		if err != nil {
+			return nil, err
+		}
+
+		// contribute the entire CR to the context
+		gvk := resource.CR.GetObjectKind().GroupVersionKind()
+		err = unstructured.SetNestedField(
+			envVarCtx, resource.CR.Object, gvk.Version, gvk.Group, gvk.Kind, resource.CR.GetName())
+		if err != nil {
+			return nil, err
+		}
+
+		// FIXME(isuttonl): make volume keys a return value
+		r.VolumeKeys = append(r.VolumeKeys, resource.VolumeKeys...)
+	}
+
+	envVarTemplates := r.plan.SBR.Spec.CustomEnvVar
+	envParser := NewCustomEnvParser(envVarTemplates, envVarCtx)
+	envVars, err := envParser.Parse()
+	if err != nil {
+		return nil, err
+	}
+
+	// convert values to a map[string][]byte
+	result := make(map[string][]byte)
+	for k, v := range envVars {
+		result[k] = []byte(v.(string))
+	}
+
+	return result, nil
+}
+
+// ReadBindableResourcesData reads all related resources of a given sbr
+func (r *Retriever) ReadBindableResourcesData(
+	sbr *v1alpha1.ServiceBindingRequest,
+	crs []*unstructured.Unstructured,
+) error {
+	r.logger.Info("Detecting extra resources for binding...")
+	for _, cr := range crs {
+		b := NewDetectBindableResources(sbr, cr, []schema.GroupVersionResource{
+			{Group: "", Version: "v1", Resource: "configmaps"},
+			{Group: "", Version: "v1", Resource: "services"},
+			{Group: "route.openshift.io", Version: "v1", Resource: "routes"},
+		}, r.client)
+
+		vals, err := b.GetBindableVariables()
+		if err != nil {
+			return err
+		}
+		for k, v := range vals {
+			r.store(cr, k, []byte(fmt.Sprintf("%v", v)))
+		}
+	}
+
+	return nil
 }
 
 // getNestedValue retrieve value from dotted key path
